Updated test files

+++ proj1/test/test_restore_8.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-TIMEOUT=60
-VERSION=1.3
+TIMEOUT=70
+VERSION=1.5
 MC_ADDR=230.0.0.1
 MC_PORT=8888
 MDB_ADDR=230.0.0.2
@@ -13,8 +13,10 @@ test () {
     echo -en "$1\t"
     expected=$($3)
     output=$($2)
-    if [ $? != 0 ]; then
-        echo -e "\e[1m\e[31m[Failed]\e[0m: return code is not zero"
+    if [ $? == 0 ] && [ "$output" == "$expected" ]; then
+        echo -e "\e[1m\e[32m[Passed]\e[0m"
+    else
+        echo -e "\e[1m\e[31m[Failed]\e[0m"
         kill $PID1
         kill $PID2
         kill $PID3
@@ -22,36 +24,32 @@ test () {
         kill $PID5
         exit 1
     fi
-    echo $expected > expected.txt
-    echo $output > output.txt
-    if ! diff expected.txt output.txt ; then
-        echo -e "\e[1m\e[31m[Failed]\e[0m: expected different from output"
-        kill $PID1
-        kill $PID2
-        kill $PID3
-        kill $PID4
-        kill $PID5
-        exit 1
-    fi
-    echo -e "\e[1m\e[32m[Passed]\e[0m"
 }
 
 cd build
-rm -rf 1 2 3 4 5
-curl http://ftp.debian.org/debian/dists/jessie/main/source/Release -o source_Release  # 102B
+rm -rf 1 2
+rm -rf testfiles
+mkdir testfiles
+curl http://ftp.debian.org/debian/dists/jessie/ChangeLog           -o testfiles/ChangeLog       # 2.3MB
 timeout $TIMEOUT java PeerDriver $VERSION 1 service1 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID1=$!
 timeout $TIMEOUT java PeerDriver $VERSION 2 service2 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID2=$!
 timeout $TIMEOUT java PeerDriver $VERSION 3 service3 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID3=$!
 timeout $TIMEOUT java PeerDriver $VERSION 4 service4 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID4=$!
-timeout $TIMEOUT java PeerDriver $VERSION 5 service5 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID5=$!
-echo "Started peers with PIDs $PID1, $PID2, $PID3, $PID4, $PID5"
+echo "Started peers with PIDs $PID1, $PID2, $PID3, $PID4"
 sleep 1
 
-test "test-backup-2-01-1" "timeout $TIMEOUT java TestApp service1 BACKUP source_Release 2" "echo"
-sleep 10
-FILE=storage/chunks/8C5A4F80497BC0C4719B9DCE7CCC75C36BCB3938A65FB65F7CC0CA0074279526
-contents=$(cat 2/$FILE-0 3/$FILE-0 4/$FILE-0 5/$FILE-0)
-test "test-backup-2-01-2" "echo $contents" "cat source_Release source_Release"
+cp testfiles/ChangeLog .
+timeout $TIMEOUT java TestApp service1 BACKUP ChangeLog 1
+sleep 40
+rm ChangeLog
+
+timeout $TIMEOUT java PeerDriver $VERSION 5 service5 $MC_ADDR $MC_PORT $MDB_ADDR $MDB_PORT $MDR_ADDR $MDR_PORT > /dev/null & PID5=$!
+echo "Started peer with PID $PID5"
+sleep 1
+
+timeout $TIMEOUT java TestApp service1 RESTORE ChangeLog
+sleep 20
+test "test-restore-8-01" "cat ChangeLog" "cat testfiles/ChangeLog"
 
 kill $PID1
 kill $PID2
